basic complete

+++ contracts/SafeMath.sol
@@ -9,7 +9,7 @@ library SafeMath {
   /**
   * @dev Multiplies two numbers, reverts on overflow.
   */
-  function mul(uint8 a, uint8 b) internal pure returns (uint8) {
+  function mul(uint256 a, uint256 b) internal pure returns (uint256) {
     // Gas optimization: this is cheaper than requiring 'a' not being zero, but the
     // benefit is lost if 'b' is also tested.
     // See: https://github.com/OpenZeppelin/openzeppelin-solidity/pull/522
@@ -17,7 +17,7 @@ library SafeMath {
       return 0;
     }
 
-    uint8 c = a * b;
+    uint256 c = a * b;
     require(c / a == b);
 
     return c;
@@ -26,9 +26,9 @@ library SafeMath {
   /**
   * @dev Integer division of two numbers truncating the quotient, reverts on division by zero.
   */
-  function div(uint8 a, uint8 b) internal pure returns (uint8) {
+  function div(uint256 a, uint256 b) internal pure returns (uint256) {
     require(b > 0); // Solidity only automatically asserts when dividing by 0
-    uint8 c = a / b;
+    uint256 c = a / b;
     // assert(a == b * c + a % b); // There is no case in which this doesn't hold
 
     return c;
@@ -37,9 +37,9 @@ library SafeMath {
   /**
   * @dev Subtracts two numbers, reverts on overflow (i.e. if subtrahend is greater than minuend).
   */
-  function sub(uint8 a, uint8 b) internal pure returns (uint8) {
+  function sub(uint256 a, uint256 b) internal pure returns (uint256) {
     require(b <= a);
-    uint8 c = a - b;
+    uint256 c = a - b;
 
     return c;
   }
@@ -47,8 +47,8 @@ library SafeMath {
   /**
   * @dev Adds two numbers, reverts on overflow.
   */
-  function add(uint8 a, uint8 b) internal pure returns (uint8) {
-    uint8 c = a + b;
+  function add(uint256 a, uint256 b) internal pure returns (uint256) {
+    uint256 c = a + b;
     require(c >= a);
 
     return c;
@@ -58,7 +58,7 @@ library SafeMath {
   * @dev Divides two numbers and returns the remainder (unsigned integer modulo),
   * reverts when dividing by zero.
   */
-  function mod(uint8 a, uint8 b) internal pure returns (uint8) {
+  function mod(uint256 a, uint256 b) internal pure returns (uint256) {
     require(b != 0);
     return a % b;
   }
